Add mutex to linenoise calls in std::rl
